Ensure all OpenAPI Values are typed (#32)
It's easy to accidentally not correctly type an endpoint.

This currently just throws an exception at runtime if an endpoint isn't correctly typed – ideally this might be done outside of runtime, but this was the lowest effort, and I really want to know as soon as I write an endpoint incorrectly that it isn't typed right.

I also fixed the one untyped path.