Infra for custom eslint rules  (#91)
This implements a rule that ensures any class ending in "DTO" only has properties which are definite or optional (end in ! or ?).

This rule doesn't make a ton of sense in isolation, it needs a companion rule which enforces that these classes have no constructors, but I'll land that in a followup.

This isn't integrated as a monorepo, partially because we haven't yet figured out our broader monorepo strategy, and partially because the way eslint depends on things is a bit different from normal typescript monorepo dependencies, so it'll take some figuring out.