Implement backend route auth (#106)
This splits the Auth Guard in two, with one guard for performing OAuth redirects, and another for validating the session for API calls.

It's a bit clunky getting the two guards to collaborate, we use a decorator read by the Session guard to turn off the Session Guard for routes where we want the OAuth redirect guard.

The frontend doesn't respond well to 401 responses from the backend at this point.