Entity and DTO cleanup (#113)
Keep entities as backend only concepts. This means we can get rid of the mikro orm shim and have a much clearer conceptual boundary between entities and DTOs, at the cost of a bit of duplication.

We do need to make sure we keep the DTOs and entities in sync, though you can see an example in ContactEntity of enforcing this.

I also renamed all "InDTO"s to "RequestDTO" and "OutDTO"s to "ResponseDTO".