A somewhat aggressive approach to type branding (#164)
Typescript has "structural" typing, which means that two types that are the same shape will be automatically converted.

Consider this example:

class Foo {}

class ADTO {
    a!: number
    fooId?: number
}

class AEntity {
    a!:number
    foo?: Foo
}

let aDTO:ADTO = {a:3};
let aEntity: AEntity = aDTO;

Typescript allows this, despite it being clearly wrong.
To avoid this, we introduce a compile time only "brand", used
only to convince typescript that these types shouldn't be castable to one another.

We also use this base type to provide common methods on DTOs and Entities, such as all Entities needing a "toDTO" method.

In addition to the regular benefits of stronger typing, this also eliminates a source of "I just wanted to add a tiny helper method, and now I need to update 40 files".
Many of our DTOs and Entities are likely to slowly deviate to the point where we won't get implicit casting, even without the brand. When that happens, you'll then need to add an explicit cast everywhere it's required, which is more of a pain to do after the fact than when you initially start using the entity and DTO.

The brand ensures we're already handling this, so you do a little bit of work up front, but don't suddenly need to add a bunch of explicit casts just because you added a helper method to a DTO.


Also, we'll need explicit casts sometimes, and it's nicer to be 100% consistent, as opposed to only explicitly casting when the set of properties diverges too far.